# api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_florinel-chis_gophercrm_internal_utils.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
      meta:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIMeta'
      success:
        type: boolean
    type: object
...
  internal_handler.DashboardStats:
    properties:
      conversion_rate:
        type: number
      open_tickets:
        type: integer
      pending_tasks:
        type: integer
      total_customers:
        type: integer
      total_leads:
        type: integer
    type: object
...
info:
  contact: {}
  description: CRM API for managing users, leads, customers, tickets, tasks and API
    keys.
  title: GopherCRM API
  version: "1.0"
paths:
...
  /dashboard/stats:
    get:
      description: 'Aggregate counts for the dashboard: total leads, total customers,
        open tickets and pending tasks, plus a conversion rate computed as customers
        / leads * 100 (0 when there are no leads). Restricted to the admin, sales
        and support roles; customer-role callers receive 403. The totals returned
        to permitted roles are unscoped and system-wide; nothing is filtered to the
        caller''s own records.'
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard statistics retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.DashboardStats'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin, sales or support role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get dashboard statistics
      tags:
      - dashboard
...
securityDefinitions:
  ApiKeyAuth:
    description: API key issued via POST /api-keys. The plaintext key is shown only
      once at creation.
    in: header
    name: ApiKey
    type: apiKey
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"